requirements.txt
pycryptodome
rsa
pyqt5